+++ db/migrate/20190130141115_create_renters.rb
@@ -7,7 +7,7 @@ class CreateRenters < ActiveRecord::Migration[5.2]
 
       t.integer :bedrooms
       t.integer :bathrooms 
-      t.integer :distance_to_subway 
+      t.float :distance_to_subway 
       t.string :borough 
       t.string :neighborhood
       t.boolean :pet_friendly
